Remove EnableOAuth21 flag
